% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BADGE.R
\name{BADGE}
\alias{BADGE}
\title{Variational Bayesian Structure Learning of Dynamic Graphical Models}
\usage{
BADGE(data, T = Inf, max_iter = 1e+06, tol_s = 0.1,
  tol_relative = 0.01, anneal_iters = 500, is_nonzero_mean = FALSE,
  normalize_data = FALSE)
}
\arguments{
\item{data}{N x P matrix of observed data. N is the sample size and P is the dimension. Missing data can be denoted by NA.}

\item{max_iter}{maximum number of iterations (1e6 by default).}

\item{tol_s}{tolerance for the maximum absolute difference between the expected state values in two consecutive iterations to check the convergence of the algorithm (1e-1 by default).}

\item{tol_relative}{tolerance for the relative difference between the estimated precision matrices in two consecutive iterations to check the convergence (1e-2 by default)}

\item{anneal_iters}{number of iterations for asymtotic simulated annealing (500 by default)}

\item{is_nonzero_mean}{boolean value to determine whether the mean of the observations is nonzero across time (FALSE by default)}

\item{normalize_data}{boolean value to determine whether to normalize the data to have unit variance across time (FALSE by default). 
It is recommended to to normalize the data before learning the time-varying graphical models}
}
\value{
EKd_mat N x P matrix of the diagonal of the precision matrices at N time points.

EKod_mat N x P(P-1)/2 matrix of the vectorized lower-triangular parts of the precision matrices at N time points

EJ_mat N x P(P-1)/2 matrix of the expectation of the vectorized lower-triangular parts of the J matrices at N time points

Es_mat N x P(P-1)/2 matrix of the expecation of the vectorized lower-triangular parts of the s matrices at N time points

data N x P matrix of the imputed data

run_time overall run time
}
\description{
Variational Bayesian Structure Learning of Dynamic Graphical Models
}
